requirements.txt
git+https://github.com/dolfies/discord.py-self.git
google-genai
python-dotenv
httpx
pillow
